% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BANEScarparking.R
\docType{package}
\name{BANEScarparking-package}
\alias{BANEScarparking-package}
\title{Car parking data for Bath and North East Somerset}
\description{
Contains datasets of records from 8 car parks located in Bath, United
 Kingdom, and some functions for working with these records. Data is
 open-source and is provided by Bath and North East Somerset Council in
 collaboration with Bath: Hacked. See \href{http://bit.ly/2i3Y1uF}{here}
 for the original dataset.
}
\author{
Owen Jones (\email{olj23@bath.ac.uk})
}

