% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_dataset}
\alias{DF}
\alias{full_dataset}
\title{Car Parking Data for Bath and North East Somerset}
\format{Data frame with 1315713 records of 7 variables:
\describe{
 \item{Name}{name of car park}
 \item{LastUpdate}{time record was taken}
 \item{DateUploaded}{time record was uploaded to server}
 \item{Occupancy}{number of cars present}
 \item{Capacity}{number of available spaces}
 \item{Status}{indicates change in occupancy since last record was taken}
 \item{Proportion}{calculated as (Occupancy/Capacity)}
}}
\source{
BANES Council, Bath: Hacked. See \url{http://bit.ly/2i3Y1uF} for the
 original dataset and \url{http://bit.ly/2i3Ub4U} for associated
 documentation.
}
\usage{
DF
}
\description{
Contains all records from 8 car parks located in Bath, United Kingdom, taken
 between October 2014 and December 2016. Data is open-source and is provided
 by Bath and North East Somerset Council in collaboration with Bath: Hacked,
 and has been processed in the way documented in \code{\link{refine}}.
}
\keyword{datasets}

