% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rugby}
\alias{rugby}
\title{Bath Rugby home matches}
\format{Data frame with 43 records of 2 variables:
\describe{
 \item{GMT}{kick-off time and date}
 \item{HomeWin}{logical indicating whether Bath were victorious}
}}
\usage{
rugby
}
\description{
Dates and kick-off times for Bath Rugby's home matches from September 2014 to
 December 2016, and the outcome of each match
}
\keyword{datasets}

